+++ pkg/gardener/structuredauth/configmap.go
@@ -1,0 +1,125 @@
+package structuredauth
+
+import (
+	"context"
+	gardener "github.com/gardener/gardener/pkg/apis/core/v1beta1"
+	v1 "k8s.io/api/core/v1"
+	k8serrors "k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/utils/ptr"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/yaml"
+)
+
+type JWTAuthenticator struct {
+	Issuer        Issuer        `json:"issuer"`
+	ClaimMappings ClaimMappings `json:"claimMappings"`
+}
+
+type Issuer struct {
+	URL       string   `json:"url"`
+	Audiences []string `json:"audiences"`
+}
+
+type ClaimMappings struct {
+	Username PrefixedClaim `json:"username"`
+	Groups   PrefixedClaim `json:"groups"`
+}
+
+type PrefixedClaim struct {
+	Claim  string  `json:"claim"`
+	Prefix *string `json:"prefix,omitempty"`
+}
+
+type AuthenticationConfiguration struct {
+	metav1.TypeMeta
+
+	JWT []JWTAuthenticator `json:"jwt"`
+}
+
+func toAuthenticationConfiguration(oidcConfig gardener.OIDCConfig) AuthenticationConfiguration {
+
+	toJWTAuthenticator := func(oidcConfig gardener.OIDCConfig) JWTAuthenticator {
+		// If Groups prefix is not set by the KEB, default is set as Gardener requires non-empty value
+		groupsPrefix := ptr.To("-")
+
+		if oidcConfig.GroupsPrefix != nil {
+			groupsPrefix = oidcConfig.GroupsPrefix
+		}
+
+		return JWTAuthenticator{
+			Issuer: Issuer{
+				URL:       *oidcConfig.IssuerURL,
+				Audiences: []string{*oidcConfig.ClientID},
+			},
+			ClaimMappings: ClaimMappings{
+				Username: PrefixedClaim{
+					Claim:  *oidcConfig.UsernameClaim,
+					Prefix: oidcConfig.UsernamePrefix,
+				},
+				Groups: PrefixedClaim{
+					Claim:  *oidcConfig.GroupsClaim,
+					Prefix: groupsPrefix,
+				},
+			},
+		}
+	}
+
+	jwtAuthenticators := make([]JWTAuthenticator, 0)
+	jwtAuthenticators = append(jwtAuthenticators, toJWTAuthenticator(oidcConfig))
+
+	return AuthenticationConfiguration{
+		TypeMeta: metav1.TypeMeta{
+			Kind:       "AuthenticationConfiguration",
+			APIVersion: "apiserver.config.k8s.io/v1beta1",
+		},
+		JWT: jwtAuthenticators,
+	}
+}
+
+func CreateOrUpdateOIDCConfigMap(ctx context.Context, shootClient client.Client, cmKey types.NamespacedName, oidcConfig gardener.OIDCConfig) error {
+	creteConfigMapObject := func() (v1.ConfigMap, error) {
+		authenticationConfig := toAuthenticationConfiguration(oidcConfig)
+		authConfigBytes, err := yaml.Marshal(authenticationConfig)
+		if err != nil {
+			return v1.ConfigMap{}, err
+		}
+
+		return v1.ConfigMap{
+			TypeMeta: metav1.TypeMeta{
+				Kind:       "ConfigMap",
+				APIVersion: "v1",
+			},
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      cmKey.Name,
+				Namespace: cmKey.Namespace,
+			},
+			Data: map[string]string{
+				"config.yaml": string(authConfigBytes),
+			},
+		}, err
+	}
+
+	var existingCM v1.ConfigMap
+	err := shootClient.Get(ctx, cmKey, &existingCM)
+
+	if err != nil && !k8serrors.IsNotFound(err) {
+		return err
+	}
+
+	configMapAlreadyExists := err == nil
+
+	newConfigMap, err := creteConfigMapObject()
+
+	if err != nil {
+		return err
+	}
+
+	if configMapAlreadyExists {
+		existingCM.Data = newConfigMap.Data
+		return shootClient.Update(ctx, &existingCM)
+	}
+
+	return shootClient.Create(ctx, &newConfigMap)
+}
